[Mod] apply autocxxpy 0.3.4b in ToraApi, XtpApi, TapApi
