+++ vnpy/api/tap/vntap/generated_files/generated_functions.h
@@ -1,5 +1,5 @@
 #pragma once
-
+#include "config.h"
 
 #include <pybind11/pybind11.h>
 #include <autocxxpy/autocxxpy.hpp>
