+++ vnpy/api/tap/vntap/generated_files/generated_functions_3.cpp
@@ -1,3 +1,5 @@
+#include "config.h"
+
 #include <iostream>
 #include <string>
 #include <pybind11/pybind11.h>
