+++ vnpy/api/tap/vntap/generated_files/wrappers.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include "config.h"
 
 #include <pybind11/pybind11.h>
 #include <autocxxpy/autocxxpy.hpp>
