+++ vnpy/api/xtp/vnxtp/generated_files/config.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#define AUTOCXXPY_ENCODING_UTF8
+
+
+
+
