+++ vnpy/api/xtp/vnxtp/generated_files/module.cpp
@@ -1,3 +1,5 @@
+#include "config.h"
+
 #include <iostream>
 #include <string>
 #include <pybind11/pybind11.h>
@@ -24,7 +26,7 @@ void init_dispatcher(pybind11::module &m)
     m.def("set_async_callback_exception_handler", &autocxxpy::async_callback_exception_handler::set_handler);
 
     pybind11::class_<autocxxpy::async_dispatch_exception> c(m, "AsyncDispatchException");
-    c.def_property("what", &autocxxpy::async_dispatch_exception::what, nullptr);
+    c.def_property_readonly("what", &autocxxpy::async_dispatch_exception::what_mutable);
     c.def_readonly("instance", &autocxxpy::async_dispatch_exception::instance);
     c.def_readonly("function_name", &autocxxpy::async_dispatch_exception::function_name);
