+++ vnpy/api/xtp/vnxtp/generated_files/module.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include "config.h"
 
 #include <autocxxpy/autocxxpy.hpp>
