+++ vnpy/api/xtp/vnxtp_XTP.pyi
@@ -7,11 +7,12 @@ if typing.TYPE_CHECKING:
     from .vnxtp import *
 
 
-def set_async_callback_exception_handler(handler: Callable[[Exception, object, str], bool]):
+def set_async_callback_exception_handler(handler: Callable[[AsyncDispatchException], None]):
     """
     set a customize exception handler for async callback in this module(pyd)
     \a handler should return True if it handles that exception,
-    If the return value of \a handler is not True, exception will be re-thrown.
+
+    :note: If the return value of \a handler is not True, exception will be re-thrown.
     """
     ...
