+++ vnpy/api/xtp/vnxtp_XTP_API.pyi
@@ -7,11 +7,12 @@ if typing.TYPE_CHECKING:
     from .vnxtp import *
 
 
-def set_async_callback_exception_handler(handler: Callable[[Exception, object, str], bool]):
+def set_async_callback_exception_handler(handler: Callable[[AsyncDispatchException], None]):
     """
     set a customize exception handler for async callback in this module(pyd)
     \a handler should return True if it handles that exception,
-    If the return value of \a handler is not True, exception will be re-thrown.
+
+    :note: If the return value of \a handler is not True, exception will be re-thrown.
     """
     ...
 
@@ -27,55 +28,55 @@ class TraderSpi():
     
     
     
-    def OnDisconnected(self, session_id: int, reason: int)->Any:
+    def OnDisconnected(self, session_id: int, reason: int)->None:
         ...
-    def OnError(self, error_info: XTPRspInfoStruct)->Any:
+    def OnError(self, error_info: XTPRspInfoStruct)->None:
         ...
-    def OnOrderEvent(self, order_info: XTPOrderInfo, error_info: XTPRspInfoStruct, session_id: int)->Any:
+    def OnOrderEvent(self, order_info: XTPOrderInfo, error_info: XTPRspInfoStruct, session_id: int)->None:
         ...
-    def OnTradeEvent(self, trade_info: XTPTradeReport, session_id: int)->Any:
+    def OnTradeEvent(self, trade_info: XTPTradeReport, session_id: int)->None:
         ...
-    def OnCancelOrderError(self, cancel_info: XTPOrderCancelInfo, error_info: XTPRspInfoStruct, session_id: int)->Any:
+    def OnCancelOrderError(self, cancel_info: XTPOrderCancelInfo, error_info: XTPRspInfoStruct, session_id: int)->None:
         ...
-    def OnQueryOrder(self, order_info: XTPOrderInfo, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->Any:
+    def OnQueryOrder(self, order_info: XTPOrderInfo, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->None:
         ...
-    def OnQueryTrade(self, trade_info: XTPTradeReport, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->Any:
+    def OnQueryTrade(self, trade_info: XTPTradeReport, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->None:
         ...
-    def OnQueryPosition(self, position: XTPQueryStkPositionRsp, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->Any:
+    def OnQueryPosition(self, position: XTPQueryStkPositionRsp, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->None:
         ...
-    def OnQueryAsset(self, asset: XTPQueryAssetRsp, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->Any:
+    def OnQueryAsset(self, asset: XTPQueryAssetRsp, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->None:
         ...
-    def OnQueryStructuredFund(self, fund_info: XTPStructuredFundInfo, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->Any:
+    def OnQueryStructuredFund(self, fund_info: XTPStructuredFundInfo, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->None:
         ...
-    def OnQueryFundTransfer(self, fund_transfer_info: XTPFundTransferNotice, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->Any:
+    def OnQueryFundTransfer(self, fund_transfer_info: XTPFundTransferNotice, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->None:
         ...
-    def OnFundTransfer(self, fund_transfer_info: XTPFundTransferNotice, error_info: XTPRspInfoStruct, session_id: int)->Any:
+    def OnFundTransfer(self, fund_transfer_info: XTPFundTransferNotice, error_info: XTPRspInfoStruct, session_id: int)->None:
         ...
-    def OnQueryETF(self, etf_info: XTPQueryETFBaseRsp, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->Any:
+    def OnQueryETF(self, etf_info: XTPQueryETFBaseRsp, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->None:
         ...
-    def OnQueryETFBasket(self, etf_component_info: XTPQueryETFComponentRsp, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->Any:
+    def OnQueryETFBasket(self, etf_component_info: XTPQueryETFComponentRsp, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->None:
         ...
-    def OnQueryIPOInfoList(self, ipo_info: XTPQueryIPOTickerRsp, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->Any:
+    def OnQueryIPOInfoList(self, ipo_info: XTPQueryIPOTickerRsp, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->None:
         ...
-    def OnQueryIPOQuotaInfo(self, quota_info: XTPQueryIPOQuotaRsp, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->Any:
+    def OnQueryIPOQuotaInfo(self, quota_info: XTPQueryIPOQuotaRsp, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->None:
         ...
-    def OnQueryOptionAuctionInfo(self, option_info: XTPQueryOptionAuctionInfoRsp, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->Any:
+    def OnQueryOptionAuctionInfo(self, option_info: XTPQueryOptionAuctionInfoRsp, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->None:
         ...
-    def OnCreditCashRepay(self, cash_repay_info: XTPCrdCashRepayRsp, error_info: XTPRspInfoStruct, session_id: int)->Any:
+    def OnCreditCashRepay(self, cash_repay_info: XTPCrdCashRepayRsp, error_info: XTPRspInfoStruct, session_id: int)->None:
         ...
-    def OnQueryCreditCashRepayInfo(self, cash_repay_info: XTPCrdCashRepayInfo, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->Any:
+    def OnQueryCreditCashRepayInfo(self, cash_repay_info: XTPCrdCashRepayInfo, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->None:
         ...
-    def OnQueryCreditFundInfo(self, fund_info: XTPCrdFundInfo, error_info: XTPRspInfoStruct, request_id: int, session_id: int)->Any:
+    def OnQueryCreditFundInfo(self, fund_info: XTPCrdFundInfo, error_info: XTPRspInfoStruct, request_id: int, session_id: int)->None:
         ...
-    def OnQueryCreditDebtInfo(self, debt_info: XTPCrdDebtInfo, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->Any:
+    def OnQueryCreditDebtInfo(self, debt_info: XTPCrdDebtInfo, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->None:
         ...
-    def OnQueryCreditTickerDebtInfo(self, debt_info: XTPCrdDebtStockInfo, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->Any:
+    def OnQueryCreditTickerDebtInfo(self, debt_info: XTPCrdDebtStockInfo, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->None:
         ...
-    def OnQueryCreditAssetDebtInfo(self, remain_amount: float, error_info: XTPRspInfoStruct, request_id: int, session_id: int)->Any:
+    def OnQueryCreditAssetDebtInfo(self, remain_amount: float, error_info: XTPRspInfoStruct, request_id: int, session_id: int)->None:
         ...
-    def OnQueryCreditTickerAssignInfo(self, assign_info: XTPClientQueryCrdPositionStkInfo, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->Any:
+    def OnQueryCreditTickerAssignInfo(self, assign_info: XTPClientQueryCrdPositionStkInfo, error_info: XTPRspInfoStruct, request_id: int, is_last: bool, session_id: int)->None:
         ...
-    def OnQueryCreditExcessStock(self, stock_info: XTPClientQueryCrdSurplusStkRspInfo, error_info: XTPRspInfoStruct, request_id: int, session_id: int)->Any:
+    def OnQueryCreditExcessStock(self, stock_info: XTPClientQueryCrdSurplusStkRspInfo, error_info: XTPRspInfoStruct, request_id: int, session_id: int)->None:
         ...
 class TraderApi():
     
@@ -85,11 +86,11 @@ class TraderApi():
     @staticmethod
     def CreateTraderApi(client_id: int, save_file_path: str, log_level: XTP_LOG_LEVEL)->XTP.API.TraderApi:
         ...
-    def Release(self, )->Any:
+    def Release(self, )->None:
         ...
     def GetTradingDay(self, )->str:
         ...
-    def RegisterSpi(self, spi: XTP.API.TraderSpi)->Any:
+    def RegisterSpi(self, spi: XTP.API.TraderSpi)->None:
         ...
     def GetApiLastError(self, )->XTPRspInfoStruct:
         ...
@@ -99,13 +100,13 @@ class TraderApi():
         ...
     def GetAccountByXTPID(self, order_xtp_id: int)->str:
         ...
-    def SubscribePublicTopic(self, resume_type: XTP_TE_RESUME_TYPE)->Any:
+    def SubscribePublicTopic(self, resume_type: XTP_TE_RESUME_TYPE)->None:
         ...
-    def SetSoftwareVersion(self, version: str)->Any:
+    def SetSoftwareVersion(self, version: str)->None:
         ...
-    def SetSoftwareKey(self, key: str)->Any:
+    def SetSoftwareKey(self, key: str)->None:
         ...
-    def SetHeartBeatInterval(self, interval: int)->Any:
+    def SetHeartBeatInterval(self, interval: int)->None:
         ...
     def Login(self, ip: str, port: int, user: str, password: str, sock_type: XTP_PROTOCOL_TYPE)->int:
         ...
@@ -164,55 +165,55 @@ class QuoteSpi():
     
     
     
-    def OnDisconnected(self, reason: int)->Any:
+    def OnDisconnected(self, reason: int)->None:
         ...
-    def OnError(self, error_info: XTPRspInfoStruct)->Any:
+    def OnError(self, error_info: XTPRspInfoStruct)->None:
         ...
-    def OnSubMarketData(self, ticker: XTPSpecificTickerStruct, error_info: XTPRspInfoStruct, is_last: bool)->Any:
+    def OnSubMarketData(self, ticker: XTPSpecificTickerStruct, error_info: XTPRspInfoStruct, is_last: bool)->None:
         ...
-    def OnUnSubMarketData(self, ticker: XTPSpecificTickerStruct, error_info: XTPRspInfoStruct, is_last: bool)->Any:
+    def OnUnSubMarketData(self, ticker: XTPSpecificTickerStruct, error_info: XTPRspInfoStruct, is_last: bool)->None:
         ...
-    def OnDepthMarketData(self, market_data: XTPMarketDataStruct, bid1_qty: List[int], bid1_count: int, max_bid1_count: int, ask1_qty: List[int], ask1_count: int, max_ask1_count: int)->Any:
+    def OnDepthMarketData(self, market_data: XTPMarketDataStruct, bid1_qty: List[int], bid1_count: int, max_bid1_count: int, ask1_qty: List[int], ask1_count: int, max_ask1_count: int)->None:
         ...
-    def OnSubOrderBook(self, ticker: XTPSpecificTickerStruct, error_info: XTPRspInfoStruct, is_last: bool)->Any:
+    def OnSubOrderBook(self, ticker: XTPSpecificTickerStruct, error_info: XTPRspInfoStruct, is_last: bool)->None:
         ...
-    def OnUnSubOrderBook(self, ticker: XTPSpecificTickerStruct, error_info: XTPRspInfoStruct, is_last: bool)->Any:
+    def OnUnSubOrderBook(self, ticker: XTPSpecificTickerStruct, error_info: XTPRspInfoStruct, is_last: bool)->None:
         ...
-    def OnOrderBook(self, order_book: OrderBookStruct)->Any:
+    def OnOrderBook(self, order_book: OrderBookStruct)->None:
         ...
-    def OnSubTickByTick(self, ticker: XTPSpecificTickerStruct, error_info: XTPRspInfoStruct, is_last: bool)->Any:
+    def OnSubTickByTick(self, ticker: XTPSpecificTickerStruct, error_info: XTPRspInfoStruct, is_last: bool)->None:
         ...
-    def OnUnSubTickByTick(self, ticker: XTPSpecificTickerStruct, error_info: XTPRspInfoStruct, is_last: bool)->Any:
+    def OnUnSubTickByTick(self, ticker: XTPSpecificTickerStruct, error_info: XTPRspInfoStruct, is_last: bool)->None:
         ...
-    def OnTickByTick(self, tbt_data: XTPTickByTickStruct)->Any:
+    def OnTickByTick(self, tbt_data: XTPTickByTickStruct)->None:
         ...
-    def OnSubscribeAllMarketData(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->Any:
+    def OnSubscribeAllMarketData(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->None:
         ...
-    def OnUnSubscribeAllMarketData(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->Any:
+    def OnUnSubscribeAllMarketData(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->None:
         ...
-    def OnSubscribeAllOrderBook(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->Any:
+    def OnSubscribeAllOrderBook(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->None:
         ...
-    def OnUnSubscribeAllOrderBook(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->Any:
+    def OnUnSubscribeAllOrderBook(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->None:
         ...
-    def OnSubscribeAllTickByTick(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->Any:
+    def OnSubscribeAllTickByTick(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->None:
         ...
-    def OnUnSubscribeAllTickByTick(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->Any:
+    def OnUnSubscribeAllTickByTick(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->None:
         ...
-    def OnQueryAllTickers(self, ticker_info: XTPQuoteStaticInfo, error_info: XTPRspInfoStruct, is_last: bool)->Any:
+    def OnQueryAllTickers(self, ticker_info: XTPQuoteStaticInfo, error_info: XTPRspInfoStruct, is_last: bool)->None:
         ...
-    def OnQueryTickersPriceInfo(self, ticker_info: XTPTickerPriceInfo, error_info: XTPRspInfoStruct, is_last: bool)->Any:
+    def OnQueryTickersPriceInfo(self, ticker_info: XTPTickerPriceInfo, error_info: XTPRspInfoStruct, is_last: bool)->None:
         ...
-    def OnSubscribeAllOptionMarketData(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->Any:
+    def OnSubscribeAllOptionMarketData(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->None:
         ...
-    def OnUnSubscribeAllOptionMarketData(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->Any:
+    def OnUnSubscribeAllOptionMarketData(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->None:
         ...
-    def OnSubscribeAllOptionOrderBook(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->Any:
+    def OnSubscribeAllOptionOrderBook(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->None:
         ...
-    def OnUnSubscribeAllOptionOrderBook(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->Any:
+    def OnUnSubscribeAllOptionOrderBook(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->None:
         ...
-    def OnSubscribeAllOptionTickByTick(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->Any:
+    def OnSubscribeAllOptionTickByTick(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->None:
         ...
-    def OnUnSubscribeAllOptionTickByTick(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->Any:
+    def OnUnSubscribeAllOptionTickByTick(self, exchange_id: XTP_EXCHANGE_TYPE, error_info: XTPRspInfoStruct)->None:
         ...
 class QuoteApi():
     
@@ -222,7 +223,7 @@ class QuoteApi():
     @staticmethod
     def CreateQuoteApi(client_id: int, save_file_path: str, log_level: XTP_LOG_LEVEL)->XTP.API.QuoteApi:
         ...
-    def Release(self, )->Any:
+    def Release(self, )->None:
         ...
     def GetTradingDay(self, )->str:
         ...
@@ -230,11 +231,11 @@ class QuoteApi():
         ...
     def GetApiLastError(self, )->XTPRspInfoStruct:
         ...
-    def SetUDPBufferSize(self, buff_size: int)->Any:
+    def SetUDPBufferSize(self, buff_size: int)->None:
         ...
-    def RegisterSpi(self, spi: XTP.API.QuoteSpi)->Any:
+    def RegisterSpi(self, spi: XTP.API.QuoteSpi)->None:
         ...
-    def SetHeartBeatInterval(self, interval: int)->Any:
+    def SetHeartBeatInterval(self, interval: int)->None:
         ...
     def SubscribeMarketData(self, ticker: List[str], exchange_id: XTP_EXCHANGE_TYPE)->int:
         ...
